refactor: modularize base64 into focused modules
Split monolithic base64.rs (1,917 lines) into modular structure:
- encoding.rs: encoding conversion helpers (301 lines)
- string_ops.rs: string-based Base64 FFI functions (256 lines)
- bytes_ops.rs: byte array-based Base64 FFI functions (156 lines)
- mod.rs: module organization and re-exports (13 lines)

Benefits:
- Improved maintainability with clear separation of concerns
- Tests co-located with implementation (Rust best practice)
- Easier navigation with smaller, focused files
- All 184 tests passing with no breaking changes
